LayoutTests/inspector/styles/styles-add-blank-property.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    WebInspector.showPanel("elements");
    InspectorTest.selectNodeAndWaitForStyles("inspected", step1);

    var treeElement;
    var section;

    function step1()
    {
        InspectorTest.addResult("Before append:");
        InspectorTest.dumpSelectedElementStyles(true);
        section = WebInspector.panels.elements.sidebarPanes.styles.sections[0][1];
        section.expand();

        // Create and increment.
        treeElement = section.addNewBlankProperty();
        treeElement.startEditing();
        treeElement.nameElement.textContent = "margin-left";
        treeElement.nameElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));

        treeElement.valueElement.textContent = "1px";
        treeElement.valueElement.firstChild.select();
        treeElement.valueElement.dispatchEvent(InspectorTest.createKeyEvent("Up"));
        InspectorTest.runAfterPendingDispatches(step2);
    }

    function step2()
    {
        // Increment again.
        treeElement.valueElement.dispatchEvent(InspectorTest.createKeyEvent("Up"));
        InspectorTest.runAfterPendingDispatches(step3);
    }

    function step3()
    {
        // Commit.
        treeElement.nameElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));
        InspectorTest.selectNodeWithId("other");
        InspectorTest.runAfterPendingDispatches(step4);
    }

    function step4()
    {
        InspectorTest.selectNodeAndWaitForStyles("inspected", step5);
    }

    function step5()
    {
        InspectorTest.addResult("After append:");
        InspectorTest.dumpSelectedElementStyles(true);

        treeElement = WebInspector.panels.elements.sidebarPanes.styles.sections[0][1].addNewBlankProperty();
        treeElement.startEditing();
        treeElement.nameElement.textContent = "color";
        treeElement.nameElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));

        treeElement.valueElement.textContent = "green; font-weight: bold";
        treeElement.kickFreeFlowStyleEditForTest();

        treeElement.valueElement.textContent = "red; font-weight: bold";
        treeElement.kickFreeFlowStyleEditForTest();

        treeElement.valueElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));
        InspectorTest.runAfterPendingDispatches(step6);
    }

    function step6()
    {
        InspectorTest.selectNodeWithId("other");
        InspectorTest.runAfterPendingDispatches(step7);
    }

    function step7()
    {
        InspectorTest.selectNodeAndWaitForStyles("inspected", step8);
    }


    function step8()
    {
        InspectorTest.addResult("After append and change of a 'compound' property:");
        InspectorTest.dumpSelectedElementStyles(true);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that adding a new blank property works.
</p>

<div id="inspected" style="font-size: 12px">Text</div>
<div id="other"></div>

</body>
</html>
